add set-operation typings tests.
