+++ test/typings/test-d/index.test-d.ts
@@ -13,22 +13,14 @@ import {
   InsertResult,
   UpdateResult,
   DeleteResult,
-  Generated,
-  GeneratedAlways,
   Selectable,
-  ColumnType,
   sql,
   ExpressionBuilder,
-} from '.'
+} from '..'
 
-import { Book, Database, Movie, Person, Pet } from './shared'
+import { Book, Database, Movie, Person, Pet } from '../shared'
 
-import {
-  expectType,
-  expectError,
-  expectAssignable,
-  expectNotAssignable,
-} from 'tsd'
+import { expectType, expectError, expectAssignable } from 'tsd'
 
 async function testFromSingle(db: Kysely<Database>) {
   // Single table
@@ -900,37 +892,6 @@ async function testExecuteTakeFirstOrThrow(db: Kysely<Database>) {
     .executeTakeFirstOrThrow()
 
   expectType<Selectable<Person>>(r1)
-}
-
-async function testUnion(db: Kysely<Database>) {
-  // Subquery
-  const r1 = await db
-    .selectFrom('person')
-    .select(['id', 'first_name as name'])
-    .where('id', 'in', [1, 2, 3])
-    .union(db.selectFrom('pet').select('name').select('owner_id as id'))
-    .executeTakeFirstOrThrow()
-
-  expectType<{ id: number; name: string }>(r1)
-
-  // Raw expression
-  const r2 = await db
-    .selectFrom('person')
-    .select(['id', 'first_name as name'])
-    .where('id', 'in', [1, 2, 3])
-    .union(sql<{ id: number; name: string }>`(1, 'Sami')`)
-    .executeTakeFirstOrThrow()
-
-  expectType<{ id: number; name: string }>(r2)
-
-  // Unioned expression has a different type
-  expectError(
-    db
-      .selectFrom('person')
-      .select(['id', 'first_name as name'])
-      .where('id', 'in', [1, 2, 3])
-      .union(db.selectFrom('pet').select('name').select('owner_id'))
-  )
 }
 
 async function testCall(db: Kysely<Database>) {
